feat: Smarter script
Does not replace the mirrorlist files if either rate-mirrors fails to run, or if the minimum number of mirrors are not found. If the minimum number of mirrors are not found, the script retries with a longer per mirror timeout. If that too fails, nothing is overwritten. Otherwise the mirrorlists are updated.